changed parameter type from int to size_t

+++ C/Playground/swap.h
@@ -1,7 +1,7 @@
 #pragma once
 //********************************************
 // char* title = "swap"
-// made by Lieman at 2020.06.04
+// made by Lieman at 2020.06.05
 //
 // description:
 //	swap example
@@ -18,12 +18,12 @@
 
 
 
-void swap(void* lhs, void* rhs, int sizeOfElement) {
+void swap(void* lhs, void* rhs, size_t sizeOfElement) {
 	unsigned char* uChLHS = (unsigned char*)lhs;
 	unsigned char* uChRHS = (unsigned char*)rhs;
 
 	unsigned char temp;
-	for (int i = 0; i < sizeOfElement; ++i) {
+	for (size_t i = 0; i < sizeOfElement; ++i) {
 		temp = uChLHS[i];
 		uChLHS[i] = uChRHS[i];
 		uChRHS[i] = temp;
